+++ Assets/Neoxider/Tests/Edit/GridSystem/Dice/GridMergeAndDiceTests.cs
@@ -149,6 +149,53 @@ namespace Neo.Editor.Tests.GridSystem
         }
 
         [Test]
+        public void DicePieceGenerator_GenerateWeighted_UsesPositiveWeightsOnly()
+        {
+            var generator = new DicePieceGenerator(_ => 0);
+
+            DicePiece piece = generator.GenerateWeighted(new[]
+            {
+                new DiceValueWeight(1, 0),
+                new DiceValueWeight(2, 10),
+                new DiceValueWeight(3, 0)
+            }, false);
+
+            Assert.That(piece.IsPair, Is.False);
+            Assert.That(piece.Cells[0].Value, Is.EqualTo(2));
+        }
+
+        [Test]
+        public void DicePieceGenerator_GenerateWeightedPair_DoesNotDuplicateValues()
+        {
+            int[] rolls = { 0, 0 };
+            int index = 0;
+            var generator = new DicePieceGenerator(max => rolls[index++ % rolls.Length] % max);
+
+            DicePiece piece = generator.GenerateWeighted(new[]
+            {
+                new DiceValueWeight(4, 10),
+                new DiceValueWeight(5, 10)
+            }, true);
+
+            Assert.That(piece.IsPair, Is.True);
+            Assert.That(piece.Cells[0].Value, Is.EqualTo(4));
+            Assert.That(piece.Cells[1].Value, Is.EqualTo(5));
+        }
+
+        [Test]
+        public void DicePieceGenerator_GenerateWeighted_RejectsMissingPositiveWeights()
+        {
+            var generator = new DicePieceGenerator();
+
+            Assert.Throws<System.ArgumentException>(() => generator.GenerateWeighted(null));
+            Assert.Throws<System.ArgumentException>(() => generator.GenerateWeighted(new[]
+            {
+                new DiceValueWeight(1, 0),
+                new DiceValueWeight(2, 0)
+            }));
+        }
+
+        [Test]
         public void DiceBoardService_IgnoresPieceWithMissingCellsWithoutThrowing()
         {
             FieldGenerator generator = CreateGenerator(2, 2);